feat: split CLI and library into separate crates
To allow for using the core business logic here outside of the CLI (eg.
integration with `-Zscript`).
