+++ Cargo.toml
@@ -1,29 +1,21 @@
-[package]
-name = "conditions"
+[workspace]
+resolver = "2"
+
+members = ["cli", "lib"]
+
+[workspace.package]
 version = "0.2.0"
 edition = "2021"
-authors = ["John Allen <johnallen3d@threedogconsulting.com>"]
-repository = "https://github.com/johnallen3d/conditions"
+authors = ["John Allen <john@threedogconsulting.com"]
+homepage = "https://github.com/johnallen3d/conditions"
+repository = "https://github.com/johnallen3d/mp-cli.git"
 description = "Fetch basic weather conditions for current or specified location"
 keywords = ["weather"]
-categories = ["command-line-utilities"]
 readme = "README.md"
-
 license = "MIT"
 
-[dependencies]
-clap = { version = "4.4.13", features = ["derive"] }
-confy = "0.5.1"
-env_logger = "0.10.1"
-eyre = "0.6.11"
-lazy_static = "1.4.0"
-log = "0.4.20"
-rustls = "0.21.9"
+[workspace.dependencies]
+eyre = "0.6.8"
 serde = { version = "1", features = ["derive"] }
-thiserror = "1.0.56"
-ureq = { version = "2.9.1", features = ["json"] }
-sqlx = { version = "0.7", features = ["macros", "runtime-tokio", "sqlite"] }
-tokio = { version = "1.34", features = ["macros"] }
-
-[dev-dependencies]
-tempfile = "3.8.1"
+thiserror = "1.0.50"
+tokio = { version = "1.33", features = ["macros"] }
