+++ cli/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "conditions-cli"
+version = { workspace = true }
+edition = { workspace = true }
+authors = { workspace = true }
+homepage = { workspace = true }
+repository = { workspace = true }
+description = "CLI tootl to fetch basic weather conditions for current or specified location"
+keywords = { workspace = true }
+categories = ["command-line-utilities", "weather"]
+readme = { workspace = true }
+license = { workspace = true }
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+clap = { version = "4.4.8", features = ["derive"] }
+conditions = { path = "../lib" }
+env_logger = "0.10.0"
+eyre = { workspace = true }
+log = "0.4.20"
+serde = { workspace = true }
+serde_json = "1.0.111"
+thiserror = { workspace = true }
+tokio = { workspace = true }
