+++ lib/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "conditions"
+version = { workspace = true }
+edition = { workspace = true }
+authors = { workspace = true }
+homepage = { workspace = true }
+repository = { workspace = true }
+description = "Fetch basic weather conditions for current or specified location"
+keywords = { workspace = true }
+categories = ["weather"]
+readme = { workspace = true }
+license = { workspace = true }
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+confy = "0.5.1"
+eyre = { workspace = true }
+lazy_static = "1.4.0"
+rustls = "0.22.1"
+sqlx = { version = "0.7", features = ["macros", "runtime-tokio", "sqlite"] }
+thiserror = { workspace = true }
+ureq = { version = "2.8.0", features = ["json"] }
+serde = { workspace = true }
+tokio = { workspace = true }
+
+[dev-dependencies]
+tempfile = "3.8.1"
