+++ rustfmt.toml
@@ -1,1 +1,3 @@
 max_width = 80
+unstable_features = true
+wrap_comments = true
